# documentation/api-docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: "1.0.0"
  description: |
      API for profession orientation and quiz-based assessment.

      ## Versioning Strategy
      This API follows URI-based semantic versioning:
      - Major versions are encoded in the base path: /api/v1.0
      - Backward-incompatible changes trigger a new major version (v2.0)
      - Backward-compatible changes are released within the same version
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:8082
  description: Generated server url
security:
- BearerAuth: []
paths:
...
  /professions/{id}:
    get:
      tags:
      - profession-controller
      summary: Get profession by ID
      description: Returns a single profession by its ID with localized fields.
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Profession not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Profession found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfessionDto"
    put:
      tags:
      - profession-controller
      summary: Update a profession
      description: Updates an existing profession. Only administrators are allowed.
      operationId: update_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessionRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Validation error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Profession not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Profession updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfessionDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfessionDto"
    delete:
      tags:
      - profession-controller
      summary: Delete a profession
      description: Deletes a profession by ID. Only administrators are allowed.
      operationId: delete_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Profession not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Profession deleted
        "403":
          description: Forbidden
...
  /categories/{id}:
    put:
      tags:
      - profession-category-controller
      summary: Update a profession category
      description: Updates an existing profession category by ID. Accessible only
        to administrators.
      operationId: update_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Validation error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Category not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Category updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfessionCategoryDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfessionCategoryDto"
    delete:
      tags:
      - profession-category-controller
      summary: Delete a profession category
      description: Deletes a profession category by ID. Accessible only to administrators.
      operationId: delete_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Category not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Category deleted
        "403":
          description: Forbidden
...
  /categories:
    get:
      tags:
      - profession-category-controller
      summary: Get all profession categories
      description: Returns a list of all profession categories. Accessible only to
        administrators.
      operationId: getAll_3
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: List of profession categories
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfessionCategoryDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfessionCategoryDto"
    post:
      tags:
      - profession-category-controller
      summary: Create a profession category
      description: Creates a new profession category. Accessible only to administrators.
      operationId: create_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Validation error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Category created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfessionCategoryDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProfessionCategoryDto"
...
components:
  schemas:
    ExceptionDto:
      type: object
      description: Standardized API error response returned for all handled exceptions
      properties:
        code:
          type: integer
          format: int32
          description: HTTP status code representing the error type
          examples:
          - 400
        time:
          type: string
          format: date-time
          description: Timestamp indicating when the error occurred (ISO-8601 format)
          examples:
          - 2025-01-15T14:32:10Z
        message:
          type: object
          description: |
            Error details. May be:
            - a string message (e.g. 'Invalid credentials')
            - a structured object containing validation errors
          examples:
          - Invalid email or password
          oneOf:
          - type: string
          - {}
...
    CreateProfessionRequest:
      type: object
      description: Request payload for creating a new profession
      properties:
        code:
          type: string
          description: Unique profession code used internally
          examples:
          - software_engineer
        title:
          type: string
          description: Human-readable profession title
          examples:
          - Software Engineer
        description:
          type: string
          description: Detailed description of the profession
          examples:
          - "Designs, builds, and maintains software systems and applications."
        mlClassCode:
          type: string
          description: ML classification code returned by the ML model
          examples:
          - SE
        categoryId:
          type: integer
          format: int32
          description: Identifier of the profession category
          examples:
          - 3
    ProfessionDto:
      type: object
      description: Represents a profession that can be recommended to users based
        on quiz results
      properties:
        id:
          type: integer
          format: int32
          description: Unique identifier of the profession
          examples:
          - 42
        code:
          type: string
          description: Short internal code of the profession
          examples:
          - software_engineer
        title:
          type: string
          description: Human-readable title of the profession
          examples:
          - Software Engineer
        description:
          type: string
          description: "Detailed description of the profession, shown to end users"
          examples:
          - "Software engineers design, develop, and maintain software systems and\
            \ applications."
        mlClassCode:
          type: string
          description: Machine Learning class code used to map ML predictions to this
            profession
          examples:
          - SE
        categoryId:
          type: integer
          format: int32
          description: Identifier of the profession category
          examples:
          - 3
...
    CreateCategoryRequest:
      type: object
      description: Request payload for creating a new profession category
      properties:
        code:
          type: string
          description: Unique code of the category
          examples:
          - IT
        name:
          type: string
          description: Human-readable category name
          examples:
          - Information Technology
        colorCode:
          type: string
          description: Hex color code associated with the category (used in UI)
          examples:
          - "#1E90FF"
          pattern: "^#[0-9A-Fa-f]{6}$"
    ProfessionCategoryDto:
      type: object
      description: Represents a category used to group professions
      properties:
        id:
          type: integer
          format: int32
          description: Unique identifier of the profession category
          examples:
          - 5
        code:
          type: string
          description: Short code identifying the category
          examples:
          - IT
        name:
          type: string
          description: Human-readable name of the category
          examples:
          - Information Technology
        colorCode:
          type: string
          description: Hex color code used for UI representation of the category
          examples:
          - "#FF8800"
          pattern: "^#[0-9A-Fa-f]{6}$"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT